requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
fastapi==0.115.12
h11==0.16.0
idna==3.10
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
spotipy==2.25.1
starlette==0.46.2
typing_extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
itsdangerous==2.2.0 # Required for SessionMiddleware
python-multipart==0.0.6  # For form handling